comic_web/requirements.txt
amqp==2.5.2
amqplib==1.0.2
anyjson==0.3.3
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
billiard==3.6.3.0
carrot==0.10.7
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
colorlog==4.1.0
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
Django==2.2.2
django-celery-results==1.2.1
django-filter==2.2.0
django-picklefield==2.1.1
django-redis==4.11.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.9
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
KindleComicConverter==5.5.2
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.5
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==7.0.0
psutil==5.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyquery==1.4.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
raven==6.10.0
redis==3.4.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
yarl==1.4.2
zipp==3.1.0
